Snaps

+++ supabase/migrations/20250212074530_create_boosts.sql
@@ -17,7 +17,7 @@ create table sniperok.user_boost (
     period smallint not null,
     user_uuid uuid not null,
     boost_type_code text not null,
-    quantity decimal(20, 8) not null default 0,
+    quantity bigint not null default 0,
     CONSTRAINT user_boost_pk
         PRIMARY KEY (period, user_uuid, boost_type_code),
     CONSTRAINT user_boost_user_fk
@@ -30,11 +30,11 @@ create table sniperok.user_boost (
 --          USER_BOOST_JOURNAL
 -- ----------------------------------------------------------------------------
 create table sniperok.user_boost_journal (
-    period smallint not null,
+    period integer not null,
     user_uuid uuid not null,
     boost_type_code text not null,
     transaction_uuid uuid not null,
-    quantity decimal(20, 8) not null default 0,
+    quantity bigint not null default 0,
     journal_timestamp timestamp with time zone not null default now(),
     reference text not null,
     CONSTRAINT user_boost_journal_pk
